update notification 大枠完成, IP表記のみ要修正
